VERA-10: Replaced `LaundryButton` and `RegenerateOutfitButton` with inlined implementations. Introduced `AddItemButton` on ClosetPage and adjusted button placement and layout for better consistency.
